it just works!

+++ emulator6502/Ram.cs
@@ -4,7 +4,7 @@ namespace emulator6502
 {
     public class Ram : Addressable
     {
-        public Ram(ushort size) : base(size)
+        public Ram(int size) : base(size)
         {
         }
     }
